Successfully ran jl5 tests with Gallifrey!
I copied the `testsjl5` from `https://github.com/polyglot-compiler/polyglot`
and changed the files to have the correct extension, and the
`pthScript` to use Gallifrey instead of jl5. I verified that the same
number of tests pass except for 2, which I'll assume is because we
actually extended JL7 instead of simply JL5.